p.234 step 1: Rename Rumbl.InfoSys to InfoSys and move from rumbl/lib/rumbl/info_sys.ex to rumbrella/apps/info_sys/lib/info_sys.ex
